feat: refactor JSX into 'components' folder for clean arch, integrate 'Add Place' functionality linked to logged-in user, embed Navbar into Maps page, and add 'constants' folder
