+++ package.json
@@ -11,13 +11,8 @@
   },
   "dependencies": {
     "@tailwindcss/vite": "^4.1.4",
-    "axios": "^1.8.4",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "react-router-dom": "^7.5.0",
-    "styled-components": "^6.1.17",
     "@uploadthing/react": "^7.3.0",
-    "axios": "^1.8.4",
+    "axios": "^1.9.0",
     "leaflet": "^1.9.4",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
