+++ src/components/LocationMarker.jsx
@@ -1,0 +1,33 @@
+import React, { useState } from 'react';
+import { Marker, useMap, useMapEvents } from 'react-leaflet';
+import L from 'leaflet';
+
+const customLocationIcon = new L.Icon({
+  iconUrl: 'icons/mylocation.png', 
+  iconSize: [80, 80], 
+});
+
+const LocationMarker = () => {
+  const [position, setPosition] = useState(null);
+  const map = useMap();
+
+  useMapEvents({
+    locationfound(e) {
+      setPosition(e.latlng);
+      map.flyTo(e.latlng, map.getZoom());
+    }
+  });
+
+  const locateUser  = () => {
+    map.locate({ setView: true, maxZoom: 16, watch: true }); 
+  };
+
+  return position === null ? null : (
+    <Marker
+      position={position}
+      icon={customLocationIcon} 
+    />
+  );
+};
+
+export default LocationMarker;
