+++ src/constants/facilities.jsx
@@ -1,0 +1,52 @@
+const FACILITIES = [
+    {
+      id: 1,
+      name: 'Jalur Kursi Roda',
+      iconSelected: "icons/jalurkursiroda-w.png",
+      iconUnselected: "icons/jalurkursiroda-b.png",
+    },
+    {
+      id: 2,
+      name: 'Pintu Otomatis',
+      iconSelected: "icons/pintuotomatis-w.png",
+      iconUnselected: "icons/pintuotomatis-b.png",
+    },
+    {
+      id: 3,
+      name: 'Parkir Disabilitas',
+      iconSelected: "icons/parkirdisabilitas-w.png",
+      iconUnselected: "icons/parkirdisabilitas-b.png",
+    },
+    {
+      id: 4,
+      name: 'Toilet Disabilitas',
+      iconSelected: "icons/toiletdisabilitas-w.png",
+      iconUnselected: "icons/toiletdisabilitas-b.png",
+    },
+    {
+      id: 5,
+      name: 'Lift Braille & Suara',
+      iconSelected: "icons/liftbraille-w.png",
+      iconUnselected: "icons/liftbraille-b.png",
+    },
+    {
+      id: 6,
+      name: 'Interpreter Isyarat',
+      iconSelected: "icons/interpreterisyarat-w.png",
+      iconUnselected: "icons/interpreterisyarat-b.png",
+    },
+    {
+      id: 7,
+      name: 'Menu Braille',
+      iconSelected: "icons/menubraille-w.png",
+      iconUnselected: "icons/menubraille-b.png",
+    },
+    {
+      id: 8,
+      name: 'Jalur Guiding Block',
+      iconSelected: "icons/jalurguildingblock-w.png",
+      iconUnselected: "icons/jalurguildingblock-b.png",
+    },
+  ];
+  
+  export default FACILITIES;
