+++ src/routes/AppRouter.jsx
@@ -7,7 +7,6 @@ import ForgotPasswordStepOne from '../pages/ForgotPasswordStepOne'
 import ForgotPasswordStepTwo from '../pages/ForgotPasswordStepTwo'
 import Dashboard from '../pages/Dashboard'
 import ChatPage from '../pages/ChatPage'
-import Maps from "../pages/Maps";
 import AddPlace from "../pages/AddPlace";
 
 export default function AppRouter() {
@@ -23,7 +22,6 @@ export default function AppRouter() {
         <Route path="/dashboard" element={<Dashboard />} />
         <Route path="/chat" element={<ChatPage />} />
         <Route path="/dashboard" element={<Dashboard />} />
-        <Route path="/maps" element={<Maps />} />
         <Route path="/tambah-tempat" element={<AddPlace />} />
       </Routes>
     </BrowserRouter>
